Rename all accessor functions to get_*(). Remove templated accessors.
This way we can always use the official type names in the function names. This
is not possible without the "get_" because some functions would otherwise have
reserved names such as "double".
Also there are no more templated functions as accessors exposed to the user.
This makes access consistent and harder to get wrong.