Add (currently undocumented) new read/write interface.
The pbf_reader and pbf_writer interface is not as typesafe as it could be.
These changes add a new thin wrapper around pbf_reader and pbf_writer called
pbf_message and pbf_builder that don't work with plain integers for tag types,
but with enum classes the user code has to define. This gives us a bit of
type checking and more self-documenting code.