Cleanly separate between reason to use bare ptr for packed_fix and use.
There are two implementations for the packed_fixed() function. One simply
returns a pair of pointers, the other a pair of iterators we define specially
for this case. The simple pointer-based implementation is presumably faster,
but it only works on little endian machines and if the CPU can work with the
possibly unaligned data.

To make this actually work on architectures that care about alignment we
still need to define PROTOZERO_USE_BARE_POINTER_FOR_PACKED_FIXED somewhere.

See #33.