Add pbf_reader::get_view() call returning object of new data_view class.
This data_view is a better way to represent a pointer to data and a
length than the std::pair used otherwise. It is compatible to the
upcoming (in C++17) string_view class. In the medium term this method
should replace the get_data(), get_string(), and get_bytes() methods
for many use cases.