Fixes Undefined Behavior due to Type Punning, resolves #64
The `byteswap` functions cast a `char *`  (alignment requirement of 1)
to a `uint32_t* ` / `uint64_t *` (higher alignment requirement).

This is Undefined Behavior.

It could e.g. fail on platforms other than x86 or if you enable
SSE and the compiler wants to make use of alignment guarantees.

The only way to mitigate this is to do a `bit_cast`-ish `memcpy`:

- http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2016/p0476r0.html
- https://github.com/jfbastien/bit_cast

The compilers are smart enough to look through all of this and will
gratefully generate a `bswap` inst. in these specific setting:

- Clang https://godbolt.org/g/NVDL31
- GCC https://godbolt.org/g/fnRheQ

```
byteswap32builtin(char const*, char*):
        movl    (%rdi), %eax
        bswap   %eax
        movl    %eax, (%rsi)
        ret

byteswap32builtinUb(char const*, char*):
        movl    (%rdi), %eax
        bswap   %eax
        movl    %eax, (%rsi)
        ret
```