Add some support for checking the wire type in addition to tags.
This adds a new version of the pbf_reader.next() function with an
additional wire_type parameter that should be checked in addition to the
tag. There is also a free function tag_and_type() and a method on
pbf_reader of the same name that can be used to check the tag and the
type in one switch statement.