Support for invalid state in pbf_writer and packed_repeated_fields.
This fixes move construction and move assignement in pbf_writer and
disables the copy construction and copy assignement which doesn't have
clear semantics. It introduces an invalid or empty state in the
pbf_writer, pbf_builder, and packed_repeated_fields classes used for
default-constructed, moved from, or committed objects. There is a new
commit() function for pbf_writer and the packed_repeated_fields which
basically does the same as the destructor but can be called explicitly.