Update clang-tidy config and add Github actions workflow
The workflow is not intended to fail when clang-tidy finds some problem.
It only helps with finding possible issues and cleaning up the code.

Clang-tidy adds lots of new checks all the time so we are running the
action on three different Debian versions which means three different
clang-tidy versions.