Fix buffer overrun for truncated bools
A truncated buffer could lead to an out of bounds memory access. Instead of checking the buffer before we are accessing the first byte beyond the end of the before, we first read it and _then_ checked it. This change reverses the actions so that no out-of-bounds read occurs.

We have not found this bug in our previous fuzzing because of the way we read the test data: std::string always adds a trailing 0 byte so that it can be trivially converted to a C string. Technically, this trailing 0 byte is part of the buffer, but is not returned as part of the buffer length. Therefore, accessing a single byte beyond the end of the buffer did not trigger the address sanitizer.